Lib/requirements.txt
svg.path
fontTools